Did A Bunch Of Stuff For The First Special Encounter Scripting, 2/26/2022
Did A Bunch Of Stuff For The First Special Encounter Scripting, 2/26/2022. Specifically involving the traveling alchemist event and adding stuff such as a trade window to be used in this event and very likely others as they get created, also new object to be attached to loot-piles/corpses for specific behaviors.